requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
poe-api==0.2.9
PySocks==1.7.1
python-socks==2.2.0
python-telegram-bot==20.2
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
socksio==1.0.0
urllib3==1.26.15
websocket-client==1.5.1
